refactor: extract core/dmxTypes.h (enum class DmxType + wire codec)
Phase 5 item 2. New Arduino-free src/core/dmxTypes.h provides
DmxType, toWireString/fromWireString (preserving the legacy
"BINARY"->Relay R5 compat alias), and the DMX_CHANNELS/PACKET_SIZE
constants. config.h's DmxType becomes a using-alias over
core::DmxType, with dmxTypeFromString/ToString as thin adapters.
artnetHandler.h's getNumberOfChannels()==512 literal now reads
core::DMX_CHANNELS. Native tests cover the enum<->string codec.

+++ src/main.cpp
@@ -185,7 +185,7 @@ void setup()
 #endif
         default:
             LOG(F("Incompatible DMX device type:"));
-            LOG(config.dmx[i]->type);
+            LOG(static_cast<int>(config.dmx[i]->type));
             break;
         }
     }
